// src/api/generated/models/book.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * API каталога книг
 * REST API для управления каталогом книг и авторов.
 * Поддерживает роли: гость (просмотр) и пользователь (CRUD).
 * Аутентификация через Bearer-токен (JWT).
 *
 * OpenAPI spec version: 1.0.0
 */
import type { AuthorShort } from './authorShort';

export interface Book {
  id?: number;
  title?: string;
  year?: number;
  description?: string;
  isbn?: string;
  cover_url?: string;
  authors?: AuthorShort[];
}