// src/api/generated/models/error.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * API каталога книг
 * REST API для управления каталогом книг и авторов.
 * Поддерживает роли: гость (просмотр) и пользователь (CRUD).
 * Аутентификация через Bearer-токен (JWT).
 *
 * OpenAPI spec version: 1.0.0
 */
import type { ErrorItem } from './errorItem';

export interface Error {
  success?: boolean;
  errors?: ErrorItem[];
}